pyproject.toml
[tool.poetry]
name = "netflix_profile"
version = "0.1.0"
description = ""
authors = ["Matheus <math.lima.andrade@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.6"
django-rest-framework = "^0.1.0"
django-restframework = "^0.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
